requirements.txt
pandas
numpy
python-dotenv
pymongo
pymongo[srv]==3.12
Certifi
-e .